refactor: cleanup old lab experiments
